server/requirements.txt
# Core
Django==5.2.7
djangorestframework==3.16.1

# Database
psycopg2-binary==2.9.11

# gRPC
grpcio==1.75.1
grpcio-tools==1.75.1

# Env management
python-dotenv==1.1.1

# Linting & formatting (optional but recommended)
black==25.9.0
isort==6.1.0
flake8==7.3.0

# Testing
pytest==8.4.2
pytest-django==4.11.1

# Pre-commit
pre-commit==4.3.0
